All events show regardless of date
All events now show regardless of the date. Previously, only events with today's date would show, however this has been changed for demo purposes. This makes it easier to show the functionality to other people in-person but also to yourself if you are reading this, as you no longer have to manually change the events dates to get them visible on the grid.